refactor: split the tags domain into the new layered architecture
* Set up the new layered architecture with api --> service --> repository
* Split the tags management domain into this new setup
* Fix a bug with deleting/renaming tags containing spaces
