+++ src/db/queries.rs
@@ -2371,34 +2371,6 @@ pub async fn delete_tags_for_link(db: &D1Database, link_id: &str) -> Result<()> 
     Ok(())
 }
 
-/// Get all tags for an org with usage counts, sorted by count desc then name asc.
-pub async fn get_org_tags(db: &D1Database, org_id: &str) -> Result<Vec<OrgTag>> {
-    let stmt = db.prepare(
-        "SELECT tag_name, COUNT(*) as count
-         FROM link_tags
-         WHERE org_id = ?1
-         GROUP BY tag_name
-         ORDER BY count DESC, tag_name ASC",
-    );
-    let results = stmt.bind(&[org_id.into()])?.all().await?;
-    let rows = results.results::<serde_json::Value>()?;
-    let tags = rows
-        .iter()
-        .filter_map(|row| {
-            let name = row["tag_name"].as_str()?.to_string();
-            let count = row["count"].as_f64()? as i64;
-            Some(OrgTag { name, count })
-        })
-        .collect();
-    Ok(tags)
-}
-
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-pub struct OrgTag {
-    pub name: String,
-    pub count: i64,
-}
-
 // ─── Org Management ───────────────────────────────────────────────────────────
 
 use crate::models::{OrgInvitation, OrgMember, OrgMemberWithUser, OrgWithRole};
@@ -3608,41 +3580,6 @@ pub async fn count_distinct_tags_for_billing_account(
     } else {
         Ok(0)
     }
-}
-
-/// Rename a tag within an organization. Updates all links that use the tag.
-/// If the new name already exists in the org, this will effectively merge the tags.
-pub async fn rename_tag_for_org(
-    db: &D1Database,
-    org_id: &str,
-    old_name: &str,
-    new_name: &str,
-) -> Result<()> {
-    // Use INSERT OR REPLACE to handle the case where new_name already exists
-    // This effectively merges the old tag into the existing one
-    let stmt = db.prepare(
-        "UPDATE link_tags
-         SET tag_name = ?1
-         WHERE org_id = ?2 AND tag_name = ?3",
-    );
-
-    stmt.bind(&[new_name.into(), org_id.into(), old_name.into()])?
-        .run()
-        .await?;
-
-    Ok(())
-}
-
-/// Delete a tag from an organization. Removes the tag from all links.
-pub async fn delete_tag_for_org(db: &D1Database, org_id: &str, tag_name: &str) -> Result<()> {
-    let stmt = db.prepare(
-        "DELETE FROM link_tags
-         WHERE org_id = ?1 AND tag_name = ?2",
-    );
-
-    stmt.bind(&[org_id.into(), tag_name.into()])?.run().await?;
-
-    Ok(())
 }
 
 // ─── Webhook Idempotency ─────────────────────────────────────────────────────
